pyproject.toml
[tool.pylint.master]
extension-pkg-whitelist = []

[tool.pylint.format]
good-names = []
max-module-lines = 4000

[tool.pylint.messages_control]
max-line-length = 120
disable = [
    'redefined-outer-name',
    'broad-exception-caught'
]

[project]
name = "simkl-data-validator"
version = "1.0.0"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
